Phase 6: reorganize daemon into src/cli/core/daemon/
Split browser-daemon.ts and daemon-ipc.ts into a dedicated daemon/
directory with focused modules:

- daemon/ipc.ts: DaemonServer, DaemonClient, protocol types, socket path
- daemon/daemon.ts: BrowserDaemon class (lifecycle, dispatch, page tracking)
- daemon/pages.ts: handlePages() extracted from BrowserDaemon
- daemon/exec.ts: handleExec(), handleReadonlyExec() extracted
- daemon/snapshot.ts: handleSnapshot() extracted
- daemon/index.ts: public API re-exports

Updated imports in execution.ts, snapshot.ts, browser.ts, exec-compiler.ts.
All 200 tests pass, type-check passes.
