+++ packages/libretto/src/cli/core/daemon/index.ts
@@ -1,0 +1,9 @@
+export {
+  DaemonServer,
+  DaemonClient,
+  getDaemonSocketPath,
+  type DaemonRequest,
+  type DaemonResponse,
+  type DaemonResultMap,
+  type RequestHandler,
+} from "./ipc.js";
